// src/app/shared/interfaces/Vehicle/vehicle.ts
export interface Vehicle {
  cargo_capacity: string;
  consumables: string;
  cost_in_credits: string;
  crew: string;
  filmIds: number[];
  films: string[];
  id: number;
  length: string;
  manufacturer: string;
  max_atmosphering_speed: string;
  model: string;
  name: string;
  passengers: string;
  pilotIds: number[];
  pilots: string[];
  url: string;
  vehicle_class: string;
}